Adds support for torsiondrive ranges that include -180
Dihedral scans support -180 to 360 degree ranges already, but giving
e.g. [-180, 0] as the range with a grid_spacing of 15 would yield a set
of optimizations starting at -165 and going up to 0. -180 would resolve
to 180, and would be outside the range given, so never computed.

Also included some readability negative space for some functions in
`dihedral_scanner`, `td_api`. Finding the root cause of this issue was
very difficult, and this was part of the process of making sense of the
existing code.